ADD route /ipcs/result.sv to route /ipcs/receipt.sv ADD response for GET for route /ipcs/result.sv

+++ main_app.py
@@ -51,30 +51,34 @@ def cut_pdf(ifile: str, ofile: str):
 
 
 # https://notes.gov4c.kz/ipsc/receipt.sv?lang=ru&appId=002224748721
+@app.route('/ipsc/result.sv', methods=['POST', 'GET'])
 @app.route('/ipsc/receipt.sv', methods=['POST', 'GET'])
 def get_request():
     status = 0
-    try:
-        lang = request.args.get('lang')
-        appId = request.args.get('appId')
-        log.info(f"SERVICE REQUESTED: appId: '{appId}', lang: {lang}")
-        if appId:
-            if not os.path.exists(f'{cfg.SPOOL}/{appId}-2.pdf'):
-                status = get_pdf(appId, lang)
+    if request.method == "POST":
+        try:
+            lang = request.args.get('lang')
+            appId = request.args.get('appId')
+            log.info(f"SERVICE REQUESTED: appId: '{appId}', lang: {lang}")
+            if appId:
+                if not os.path.exists(f'{cfg.SPOOL}/{appId}-2.pdf'):
+                    status = get_pdf(appId, lang)
+                else:
+                    status = 2
+        except Exception as e:
+            log.error(f"GET VALUES: {appId} : {lang}. error: {e}")
+        finally:
+            if status > 0:
+                if status == 1:
+                    cut_pdf(f"{cfg.SPOOL}/{appId}.pdf", f"{cfg.SPOOL}/{appId}-2.pdf")
+                    os.remove(f"{cfg.SPOOL}/{appId}.pdf")
+                log.info(f"Передаем расписку №{appId}")
+                return send_from_directory(f"{cfg.SPOOL}", f"{appId}-2.pdf")
             else:
-                status = 2
-    except Exception as e:
-        log.error(f"GET VALUES: {appId} : {lang}. error: {e}")
-    finally:
-        if status > 0:
-            if status == 1:
-                cut_pdf(f"{cfg.SPOOL}/{appId}.pdf", f"{cfg.SPOOL}/{appId}-2.pdf")
-                os.remove(f"{cfg.SPOOL}/{appId}.pdf")
-            log.info(f"Передаем расписку №{appId}")
-            return send_from_directory(f"{cfg.SPOOL}", f"{appId}-2.pdf")
-        else:
-            return f"<html><h1>Запрошенная расписка №'{appId} не найдена'</h1></html>"
-
+                return f"<html><h1>Запрошенная расписка №'{appId} не найдена'</h1></html>", 400
+    res = make_response(f'<h2>Service READY.</h2><p>Must be used POST method like this:<p><h3>'
+                        f'https://notes.gov4c.kz/ipsc/receipt.sv?lang=ru&appId=002228892389</h3')
+    return res, 200
 
 # https://notes.gov4c.kz/ipsc/receipt.sv?lang=ru&appId=002224748721
 @app.route('/')
